+++ wcp-tip-calculator/Extensions/Color+.swift
@@ -1,0 +1,8 @@
+//
+//  Color+.swift
+//  wcp-tip-calculator
+//
+//  Created by Eden Momchilov on 1/7/24.
+//
+
+import Foundation
