+++ wcp-tip-calculator/Extensions/Float+.swift
@@ -1,0 +1,8 @@
+//
+//  Float+.swift
+//  wcp-tip-calculator
+//
+//  Created by Eden Momchilov on 1/7/24.
+//
+
+import Foundation
