+++ wcp-tip-calculator/Extensions/String+.swift
@@ -1,0 +1,8 @@
+//
+//  String+.swift
+//  wcp-tip-calculator
+//
+//  Created by Eden Momchilov on 1/7/24.
+//
+
+import Foundation
