+++ wcp-tip-calculator/Models/Server.swift
@@ -1,0 +1,8 @@
+//
+//  Server.swift
+//  wcp-tip-calculator
+//
+//  Created by Eden Momchilov on 1/7/24.
+//
+
+import Foundation
