+++ wcp-tip-calculator/Views/ContentView.swift
