--- wcp-tip-calculatorTests/wcp_tip_calculatorTests.swift
@@ -1,33 +1,0 @@
-//
-//  wcp_tip_calculatorTests.swift
-//  wcp-tip-calculatorTests
-//
-//  Created by Eden  Momchilov on 10/22/21.
-//
-
-import XCTest
-@testable import wcp_tip_calculator
-
-class wcp_tip_calculatorTests: XCTestCase {
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
